requirements.txt
numpy~=1.18.5
opencv-python~=4.3.0.36
pillow~=8.4.0
matplotlib~=3.5.0
torchvision~=0.8.2